+++ app/graphql/resolvers/rdf.ts
@@ -20,14 +20,13 @@ import {
 } from "@/graphql/resolver-types";
 import { resolveDimensionType, resolveMeasureType } from "@/graphql/resolvers";
 import { defaultLocale } from "@/locales/locales";
-import { parseIri } from "@/rdf/parse";
+import { parseCube, parseIri } from "@/rdf/parse";
 import {
   createCubeDimensionValuesLoader,
   getCubeDimensions,
   getCubeObservations,
   getCubesDimensions,
   getLatestCube,
-  getResolvedCube,
 } from "@/rdf/queries";
 import { unversionObservation } from "@/rdf/query-dimension-values";
 import { queryHierarchy } from "@/rdf/query-hierarchies";
@@ -83,21 +82,27 @@ export const searchCubes: NonNullable<QueryResolvers["searchCubes"]> = async (
 };
 
 export const dataCubeByIri: NonNullable<QueryResolvers["dataCubeByIri"]> =
-  async (_, { iri, locale, latest }, { setup }, info) => {
+  async (_, { iri, locale, latest = true }, { setup }, info) => {
     const { loaders } = await setup(info);
-    const cube = await loaders.cube.load(iri);
+    const rawCube = await loaders.cube.load(iri);
 
-    if (!cube) {
+    if (!rawCube) {
       throw new Error("Cube not found");
     }
 
-    return getResolvedCube({ cube, locale: locale ?? defaultLocale, latest });
+    const cube = latest ? await getLatestCube(rawCube) : rawCube;
+    await cube.fetchShape();
+
+    return parseCube({ cube, locale: locale ?? defaultLocale });
   };
 
 export const possibleFilters: NonNullable<QueryResolvers["possibleFilters"]> =
   async (_, { iri, filters }, { setup }, info) => {
     const { sparqlClient, loaders, cache } = await setup(info);
-    const cube = await loaders.cube.load(iri);
+    const rawCube = await loaders.cube.load(iri);
+    // Currently we always default to the latest cube.
+    const cube = await getLatestCube(rawCube);
+    await cube.fetchShape();
 
     if (!cube) {
       return [];
@@ -336,15 +341,13 @@ export const hierarchy: NonNullable<DimensionResolvers["hierarchy"]> = async (
     throw new Error("Could not find cube");
   }
 
-  const res = await queryHierarchy(
+  return await queryHierarchy(
     rdimension,
     locale,
     sparqlClient,
     sparqlClientStream,
     cache
   );
-
-  return res;
 };
 
 export const dimensionValues: NonNullable<