+++ app/rdf/create-source.ts
@@ -1,13 +1,14 @@
 import { Source } from "rdf-cube-view-query";
 import rdf from "rdf-ext";
+import ParsingClient from "sparql-http-client/ParsingClient";
 
 export const pragmas = `#pragma describe.strategy cbd
 #pragma join.hash off
 `;
 
-export const createSource = ({ endpointUrl }: { endpointUrl: string }) => {
+export const createSource = (sparqlClient: ParsingClient) => {
   return new Source({
-    endpointUrl,
+    client: sparqlClient,
     queryOperation: "postUrlencoded",
     queryPrefix: pragmas,
     sourceGraph: rdf.defaultGraph(),