+++ app/rdf/query-hierarchies.ts
@@ -1,11 +1,10 @@
 import {
-  getHierarchy,
   HierarchyNode,
+  getHierarchy,
 } from "@zazuko/cube-hierarchy-query/index";
 import { AnyPointer } from "clownface";
 import orderBy from "lodash/orderBy";
 import uniqBy from "lodash/uniqBy";
-import { Cube } from "rdf-cube-view-query";
 import rdf from "rdf-ext";
 import { StreamClient } from "sparql-http-client";
 import { ParsingClient } from "sparql-http-client/ParsingClient";
@@ -15,6 +14,7 @@ import { parseTerm } from "@/domain/data";
 import { truthy } from "@/domain/types";
 import { HierarchyValue } from "@/graphql/resolver-types";
 import { ResolvedDimension } from "@/graphql/shared-types";
+import { ExtendedCube } from "@/rdf/extended-cube";
 
 import * as ns from "./namespace";
 import { getCubeDimensionValuesWithMetadata } from "./queries";
@@ -62,9 +62,9 @@ const toTree = (
               .filter((d) => d.label)
           ),
           position: parseTerm(node.resource.out(ns.schema.position).term),
-          identifier: identifier,
+          identifier,
           depth,
-          dimensionIri: dimensionIri,
+          dimensionIri,
         }
       : undefined;
     return res;
@@ -73,37 +73,29 @@ const toTree = (
   return sortChildren(results.map((r) => serializeNode(r, 0)).filter(truthy));
 };
 
-const findHierarchiesForDimension = (
-  cube: Cube,
+const getDimensionHierarchies = async (
+  cube: ExtendedCube,
   dimensionIri: string,
   locale: string
 ) => {
-  const newHierarchies = uniqBy(
-    cube.ptr
-      .any()
-      .has(ns.sh.path, rdf.namedNode(dimensionIri))
-      .has(ns.cubeMeta.inHierarchy)
-      .out(ns.cubeMeta.inHierarchy)
-      .toArray(),
-    (x) => getName(x, locale)
-  );
+  const cubeIri = cube.term?.value;
 
-  if (newHierarchies) {
-    return newHierarchies;
+  if (!cubeIri) {
+    throw new Error("Cube must have an iri!");
   }
 
-  const legacyHierarchies = cube.ptr
+  // Can't rely on the cube here, because sometimes it might contain duplicate
+  // hierarchies, due to fetching both cube and shape when cube is initialized.
+  // Here we are only interested in shape hierarchies, that's why we don't have
+  // to initialize the cube, but rather just fetch its shape.
+  const hierarchies = cube.shapePtr
     .any()
     .has(ns.sh.path, rdf.namedNode(dimensionIri))
-    .has(ns.cubeMeta.hasHierarchy)
-    .out(ns.cubeMeta.hasHierarchy)
+    .has(ns.cubeMeta.inHierarchy)
+    .out(ns.cubeMeta.inHierarchy)
     .toArray();
 
-  if (legacyHierarchies) {
-    return legacyHierarchies;
-  }
-
-  return [];
+  return uniqBy(hierarchies, (hierarchy) => getName(hierarchy, locale));
 };
 
 export const queryHierarchy = async (
@@ -113,13 +105,13 @@ export const queryHierarchy = async (
   sparqlClientStream: StreamClient,
   cache: LRUCache | undefined
 ): Promise<HierarchyValue[] | null> => {
-  const hierarchies = findHierarchiesForDimension(
-    rdimension.cube,
-    rdimension.data.iri,
-    locale
-  );
+  const {
+    cube,
+    data: { iri },
+  } = rdimension;
+  const hierarchyPointers = await getDimensionHierarchies(cube, iri, locale);
 
-  if (hierarchies.length === 0) {
+  if (hierarchyPointers.length === 0) {
     return null;
   }
 
@@ -131,25 +123,25 @@ export const queryHierarchy = async (
     cache,
   });
 
-  const allHierarchies = await Promise.all(
-    hierarchies?.map(async (h) => ({
-      // @ts-ignore
-      nodes: (await getHierarchy(h).execute(sparqlClientStream, rdf)) || [],
-      hierarchyName: getName(h.out(ns.cubeMeta.nextInHierarchy), locale),
-    }))
+  const hierarchies = await Promise.all(
+    hierarchyPointers.map(async (pointer) => {
+      return {
+        // @ts-ignore
+        nodes: await getHierarchy(pointer).execute(sparqlClientStream, rdf),
+        hierarchyName: getName(pointer, locale),
+      };
+    })
   );
 
-  const trees = allHierarchies.map((h) => {
-    const tree: (HierarchyValue & { hierarchyName?: string })[] = toTree(
-      h.nodes,
-      rdimension.data.iri,
-      locale
-    );
+  const trees = hierarchies.map(({ nodes, hierarchyName }) => {
+    const tree: (HierarchyValue & {
+      hierarchyName?: string;
+    })[] = toTree(nodes, iri, locale);
 
     if (tree.length > 0) {
       // Augment hierarchy value with hierarchyName so that when regrouping
       // below, we can create the fake nodes
-      tree[0].hierarchyName = h.hierarchyName;
+      tree[0].hierarchyName = hierarchyName;
     }
     return tree;
   });