+++ app/test/cube-view-query-mock.ts
@@ -1,0 +1,11 @@
+export const RDFCubeViewQueryMock = jest.mock("rdf-cube-view-query", () => ({
+  Node: class {
+    constructor() {}
+  },
+  Source: class {
+    constructor() {}
+  },
+  Cube: class {
+    constructor() {}
+  },
+}));