+++ app/typings/rdf.d.ts
@@ -1,8 +1,10 @@
 declare module "rdf-cube-view-query" {
   import { AnyPointer, ClownfaceInit } from "clownface";
+  import DatasetExt from "rdf-ext/lib/Dataset";
   import DefaultGraphExt from "rdf-ext/lib/DefaultGraph";
-  import { Literal, NamedNode, Term } from "rdf-js";
+  import { Literal, NamedNode, Quad, Term } from "rdf-js";
   import { ParsingClient } from "sparql-http-client/ParsingClient";
+
   type NodeInit = {
     parent?: Node;
   } & ClownfaceInit;
@@ -13,11 +15,17 @@ declare module "rdf-cube-view-query" {
     term: AnyPointer["term"];
     out: AnyPointer["out"];
     in: AnyPointer["in"];
-    dataset: AnyPointer["dataset"];
+    dataset: DatasetExt;
     clear(): void;
   }
 
+  export type CubeOptions = NodeInit & {
+    source?: Source;
+    term?: Term;
+  };
+
   export class Cube extends Node {
+    constructor(options: CubeOptions);
     static filter: {
       isPartOf: (container: $FixMe) => $FixMe;
       noValidThrough: () => $FixMe;
@@ -28,7 +36,12 @@ declare module "rdf-cube-view-query" {
     };
     dimensions: CubeDimension[];
     source: CubeSource;
+    cubeQuery(): string;
+    shapeQuery(): string;
+    async fetchCube(): Promise<void>;
     async fetchShape(): Promise<void>;
+    async init(): Promise<void>;
+    quads: Quad[];
   }
 
   export class CubeDimension {
@@ -88,6 +101,7 @@ declare module "rdf-cube-view-query" {
     dimension(options: { cubeDimension: NamedNode | string }): Dimension | null;
     observationsQuery({ disableDistinct }: { disableDistinct?: boolean }): {
       query: $FixMe;
+      previewQuery: $FixMe;
       dimensionMap: Map;
     };
     async observations({
@@ -95,21 +109,32 @@ declare module "rdf-cube-view-query" {
     }: {
       disableDistinct?: boolean;
     }): Promise<Record<string, Literal | NamedNode>[]>;
+    async preview(options: {
+      limit?: number;
+      offset?: number;
+    }): Promise<Record<string, Literal | NamedNode>[]>;
     addDimension(dimension: Dimension): View;
     createDimension(options: $FixMe): Dimension;
     setDefaultColumns(): void;
   }
-  export class Source extends Node {
-    constructor(
-      options: NodeInit & {
-        endpointUrl: string;
-        sourceGraph?: string | DefaultGraphExt;
-        user?: string;
-        password?: string;
-        queryOperation?: "get" | "postUrlencoded" | "postDirect";
-        queryPrefix?: string;
-      }
+
+  export type SourceOptions = NodeInit & {
+    sourceGraph?: string | DefaultGraphExt;
+    queryOperation?: "get" | "postUrlencoded" | "postDirect";
+    queryPrefix?: string;
+    term?: Term;
+  } & (
+      | {
+          endpointUrl: string;
+          user?: string;
+          password?: string;
+        }
+      | {
+          client?: ParsingClient;
+        }
     );
+  export class Source extends Node {
+    constructor(options: SourceOptions);
     async cube(term: Term | string): Promise<Cube | null>;
     endpoint: string;
     async cubes(options?: {
@@ -120,9 +145,11 @@ declare module "rdf-cube-view-query" {
     client: ParsingClient;
     queryOperation?: "get" | "postUrlencoded" | "postDirect";
   }
+
   export class LookupSource extends Source {
     static fromSource(source: Source): LookupSource;
     queryPrefix?: string;
   }
+
   export class CubeSource extends Source {}
 }